+++ tests/sql_test_files/it_works_set_variable.sql
@@ -1,2 +1,7 @@
-set what_does_it_do = 'wo' || 'rks';
-select 'text' as component, 'It ' || $what_does_it_do || ' !' as contents;
+set url = sqlpage.set_variable('y', '2');
+set path = sqlpage.path();
+select 'text' as component,
+    case
+        when $url = $path || '?x=1&y=2' OR $url = $path || '?y=2&x=1' THEN 'It works !'
+        else 'error: ' || $url
+    end as contents;