+++ tests/sql_test_files/it_works_set_variable_replace.sql
@@ -1,0 +1,7 @@
+set url = sqlpage.set_variable('x', '2');
+set path = sqlpage.path();
+select 'text' as component,
+    case
+        when $url = $path || '?x=2' THEN 'It works !'
+        else 'error: ' || $url
+    end as contents;